Using Threads to create GUIs
ManyPanels: a GUI that displays 16 gray panels that has a black line
that starts at the top and moves to the bottom of the panel.

OnePanel: a GUI that displays a single gray panel that has a black line
that starts at the top and moves to the bottom of the panel.

OneRing:  a GUI that displays a black panel that has a red ring that
blinks when start button is pressed.

TwoRings:  GUI that displays a two black panels. One that has a red
ring another that has a blue ring. Both rings start blinking when start
button is pressed.